LPD-77576 Extract util to reuse
